Reorganize deps
